pyproject.toml
[project]
name = "ddd-kafka-chat"
version = "0.1.0"
description = ""
authors = [
    {name = "antonnagulin",email = "nagulin03@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "motor (>=3.7.1,<4.0.0)",
    "uvicorn[all] (>=0.41.0,<0.42.0)",
    "ipython (>=9.11.0,<10.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "ruff (>=0.15.6,<0.16.0)",
    "black (>=26.3.1,<27.0.0)",
    "faker (>=40.11.1,<41.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "punq (>=0.7.0,<0.8.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "aiokafka (>=0.13.0,<0.14.0)",
    "orjson (>=3.11.8,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

# Проверяем только ошибки PEP8, связанные с комментариями
lint.select = [
    "E",   # базовые ошибки PEP8 (включает E262, E265 для комментариев)
    "F",   # ошибки кода (undefined, unused)
    "I",   # сортировка импортов
    "DJ",  # Django-правила
]

# Игнорируем проверки docstring
lint.ignore = [
    "D100",  # Missing docstring in module
    "D101",  # Missing docstring in class
    "D102",  # Missing docstring in __init__
    "D103",  # Missing docstring in public function
    "D104",  # Missing docstring in public method
    "D105",  # Missing docstring in magic method
    "D106",  # Missing docstring in public nested class
    "D107",  # Missing docstring in __init__ method
]

# Папки, которые не трогаем
exclude = [
    ".venv",
    "migrations",
    "__pycache__",
    "settings.py",
]